shopinvader_portal_mode: split common test class
